requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.7.0
huggingface-hub==0.19.4
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
numpy==1.24.3
